+++ src/config.js
@@ -1,13 +1,12 @@
 // PARAMS
 export const CHAOTIC_MAP_TYPE = 'chebyshev';
 export const CHAOTIC_MAP_DIMENSION = 1;
-export const NUM_FRUIT_FLIES = 30;
-export const NUM_ITERATIONS = 300;
+export const NUM_FRUIT_FLIES = 100;
+export const NUM_ITERATIONS = 700;
 export const NUM_TRIALS = 50;
-export const WIDTH_LOWER_BOUND = 0;
-export const WIDTH_UPPER_BOUND = 100;
+export const WIDTH_LOWER_BOUND = -10;
+export const WIDTH_UPPER_BOUND = 10;
 
 // MISC
-export const DISTANCE_TO_FOOD_DEFAULT = 0;
 export const ROOT_POWER = 0.5;
 export const SQUARED_POWER = 2;