+++ src/print.js
@@ -1,4 +1,4 @@
-import { NUM_FRUIT_FLIES, NUM_ITERATIONS } from './config';
+import { NUM_FRUIT_FLIES, NUM_ITERATIONS, NUM_TRIALS } from './config';
 
 const el = document.getElementById('results');
 
@@ -22,9 +22,71 @@ export const printCsvBestPositionPerTrial = (trialResults, element = el) => {
   });
 };
 
-export const printCsvFoodSwarmCoordinatesPerTrial = (trialResults, element = el) => {
+export const printCsvFruitFlyCoordinatesPerSingleTrial = (trialResults, element = el) => { 
   element.append('\n\n');
-  element.append('CSV: food and swarm positions per trial');
+  element.append('CSV: fruit fly coordinates per single trial');
+  element.append('\n\n');
+  element.append('trial_index');
+  element.append(',');
+  element.append('iteration_index');
+  element.append(',');
+  element.append('fruit_fly_index');
+  element.append(',');
+  element.append('fruit_fly_coordinate_x');
+  element.append(',');
+  element.append('fruit_fly_coordinate_y');
+
+  const fruitFlyIndex = Math.floor(Math.random() * NUM_FRUIT_FLIES);
+  const trialIndex = Math.floor(Math.random() * NUM_TRIALS);
+  const trial = trialResults[trialIndex];
+
+  trial.forEach((iterationResults, iterationIndex) => {
+    const { vision } = iterationResults;
+    const { fruitFlies } = vision;
+    const { fruitFlies: fruitFlyInstances } = fruitFlies;
+
+    element.append('\n');
+    element.append(trialIndex);
+    element.append(',');
+    element.append(iterationIndex);
+    element.append(',');
+    element.append(fruitFlyIndex);
+    element.append(',');
+    element.append(fruitFlyInstances[fruitFlyIndex].coordinates.x);
+    element.append(',');
+    element.append(fruitFlyInstances[fruitFlyIndex].coordinates.y);
+  });
+};
+
+export const printCsvSmellConcentrationPerSingleTrial = (trialResults, element = el) => { 
+  element.append('\n\n');
+  element.append('CSV: smell concentration per single trial');
+  element.append('\n\n');
+  element.append('trial_num');
+  element.append(',');
+  element.append('iteration_num');
+  element.append(',');
+  element.append('best_position_smell_concentration');
+
+  const trialIndex = Math.floor(Math.random() * NUM_TRIALS);
+  const trial = trialResults[trialIndex];
+
+  trial.forEach((iterationResults, iterationIndex) => {
+    const { vision } = iterationResults;
+    const { bestPosition } = vision;
+
+    element.append('\n');
+    element.append(trialIndex);
+    element.append(',');
+    element.append(iterationIndex);
+    element.append(',');
+    element.append(bestPosition.smellConcentration);
+  });
+};
+
+export const printCsvFoodSwarmCoordinatesPerSingleTrial = (trialResults, element = el) => { 
+  element.append('\n\n');
+  element.append('CSV: food and swarm coordinates per single trial');
   element.append('\n\n');
   element.append('trial_num');
   element.append(',');
@@ -33,12 +95,12 @@ export const printCsvFoodSwarmCoordinatesPerTrial = (trialResults, element = el)
   element.append('food_coordinate_y');
   element.append(',');
   element.append('swarm_coordinate_x');
-  element.append(',');
-  element.append('swarm_coordinate_y');
 
-  trialResults.forEach((trialResult, trailIndex) => {
-    const finalIteration = trialResult[NUM_ITERATIONS - 1];
-    const { vision } = finalIteration;
+  const trialIndex = Math.floor(Math.random() * NUM_TRIALS);
+  const trial = trialResults[trialIndex];
+
+  trial.forEach((iterationResults, iterationIndex) => {
+    const { vision } = iterationResults;
     const { food, swarm } = vision;
 
     const { coordinates: coordinatesSwarm } = swarm;
@@ -48,7 +110,9 @@ export const printCsvFoodSwarmCoordinatesPerTrial = (trialResults, element = el)
     const { x: xFood, y: yFood } = coordinatesFood;
     
     element.append('\n');
-    element.append(trailIndex);
+    element.append(trialIndex);
+    element.append(',');
+    element.append(iterationIndex);
     element.append(',');
     element.append(xFood);
     element.append(',');